# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Union Bank API",
    "description": "REST API for the Union Bank Management System.\n\n**API Versions**\n- `/api/v1/` \u2014 Legacy endpoints (bare response models, backward compatible)\n- `/api/v2/` \u2014 Current endpoints with standardised `ApiResponse[T]` envelope\n\nAll endpoints return JSON. Authentication uses Bearer JWT tokens.\nUse `/docs` for interactive API documentation.",
    "termsOfService": "https://union-bank.example.com/terms",
    "contact": {
      "name": "Union Bank Dev Team"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v2/savings": {
      "get": {
        "summary": "V2 List Savings Goals",
        "description": "List all savings goals for the authenticated customer.",
        "operationId": "v2_list_savings_goals_api_v2_savings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SavingsGoalsSummary_"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "V2 Create Savings Goal",
        "description": "Create a new savings goal.",
        "operationId": "v2_create_savings_goal_api_v2_savings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingsGoalCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SavingsGoalOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/savings/{goal_id}/contribute": {
      "post": {
        "summary": "V2 Contribute To Goal",
        "description": "Contribute money from your balance to a savings goal.",
        "operationId": "v2_contribute_to_goal_api_v2_savings__goal_id__contribute_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Goal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingsGoalContribute"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SavingsGoalOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/savings/{goal_id}": {
      "delete": {
        "summary": "V2 Delete Savings Goal",
        "description": "Delete a savings goal and refund the amount to your balance.",
        "operationId": "v2_delete_savings_goal_api_v2_savings__goal_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Goal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MessageData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/savings": {
      "get": {
        "summary": "List Savings Goals",
        "description": "List all savings goals for the authenticated customer.",
        "operationId": "list_savings_goals_api_savings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingsGoalsSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "Create Savings Goal",
        "description": "Create a new savings goal.",
        "operationId": "create_savings_goal_api_savings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingsGoalCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingsGoalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/savings/{goal_id}": {
      "put": {
        "summary": "Update Savings Goal",
        "description": "Update a savings goal.",
        "operationId": "update_savings_goal_api_savings__goal_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Goal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingsGoalUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingsGoalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Savings Goal",
        "description": "Delete a savings goal and refund the amount to your balance.",
        "operationId": "delete_savings_goal_api_savings__goal_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Goal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/savings/{goal_id}/contribute": {
      "post": {
        "summary": "Contribute To Goal",
        "description": "Contribute money from your balance to a savings goal.",
        "operationId": "contribute_to_goal_api_savings__goal_id__contribute_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Goal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingsGoalContribute"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingsGoalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_MessageData_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MessageData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "title": "ApiResponse[MessageData]"
      },
...
      "ApiResponse_SavingsGoalOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SavingsGoalOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "title": "ApiResponse[SavingsGoalOut]"
      },
      "ApiResponse_SavingsGoalsSummary_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SavingsGoalsSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "title": "ApiResponse[SavingsGoalsSummary]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageData": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageData"
      },
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "success"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "SavingsGoalContribute": {
        "properties": {
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "description": "Amount to contribute"
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "SavingsGoalContribute"
      },
      "SavingsGoalCreate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "title": "Name",
            "description": "Goal name"
          },
          "target_amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Target Amount",
            "description": "Savings target"
          },
          "target_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date",
            "description": "Optional target date (YYYY-MM-DD)"
          }
        },
        "type": "object",
        "required": [
          "name",
          "target_amount"
        ],
        "title": "SavingsGoalCreate"
      },
      "SavingsGoalOut": {
        "properties": {
          "goal_id": {
            "type": "string",
            "title": "Goal Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "target_amount": {
            "type": "number",
            "title": "Target Amount"
          },
          "current_amount": {
            "type": "number",
            "title": "Current Amount"
          },
          "target_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "is_completed": {
            "type": "boolean",
            "title": "Is Completed"
          },
          "progress_pct": {
            "type": "number",
            "title": "Progress Pct",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "goal_id",
          "name",
          "target_amount",
          "current_amount",
          "created_at",
          "is_completed"
        ],
        "title": "SavingsGoalOut"
      },
      "SavingsGoalUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "target_amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Amount"
          },
          "target_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date"
          }
        },
        "type": "object",
        "title": "SavingsGoalUpdate"
      },
      "SavingsGoalsSummary": {
        "properties": {
          "total_goals": {
            "type": "integer",
            "title": "Total Goals"
          },
          "completed": {
            "type": "integer",
            "title": "Completed"
          },
          "total_saved": {
            "type": "number",
            "title": "Total Saved"
          },
          "total_saved_formatted": {
            "type": "string",
            "title": "Total Saved Formatted"
          },
          "total_target": {
            "type": "number",
            "title": "Total Target"
          },
          "total_target_formatted": {
            "type": "string",
            "title": "Total Target Formatted"
          },
          "goals": {
            "items": {
              "$ref": "#/components/schemas/SavingsGoalOut"
            },
            "type": "array",
            "title": "Goals"
          }
        },
        "type": "object",
        "required": [
          "total_goals",
          "completed",
          "total_saved",
          "total_saved_formatted",
          "total_target",
          "total_target_formatted",
          "goals"
        ],
        "title": "SavingsGoalsSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}